backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.1
pydantic-settings==2.6.1
python-dotenv==1.0.1
python-multipart==0.0.17
Pillow==11.0.0